Lab 07 - Text Files
In lab 7 we learned how to read and write files.

numProcess.py - a simple program (numProcess.py) to read a collection of integers from a text file and produce a printed summary.
lineNumbers.py - A program that reads a file and output the content as numbered lines and paragraphs. The line numbers are formatted such that they are  right-justified in a field 5 characters wide, followed by a period, then a space, then the line of text.